requirements.txt
 weaviate-client
 langchain
 openai
 langchain_openai
 python-dotenv